carbon-tracker-backend/requirements.txt
Django==5.2.6
ibm_watson_machine_learning==1.0.368
python-decouple==3.8
whitenoise[brotli]==6.6.0
gunicorn==22.0.0
django-cors-headers
djangorestframework
ibmcloudant==0.9.1
ibm-watson

